+++ deltafi-ui/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Limit ingress flow selection on Upload Files page to running ingress flows (#204)
 - Reworked plugins page to make editing variables safer and more user-friendly (#203)
 
 ## [0.96.4]
@@ -316,7 +317,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - NiFi link from Dashboard
 - Requested resource information from System Metrics
 
-[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.4...main
+[unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.4...main
 [0.96.4]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.3...0.96.4
 [0.96.3]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.2...0.96.3
 [0.96.2]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.96.1...0.96.2